refactor: move kontakt/actions.ts to lib/inquiry/actions.ts
